+++ docs/release-notes.md
@@ -21,6 +21,7 @@ nav_order: 9
   flatcar 1.1.0-exp, openshift 4.14.0-exp)_
 - Allow specifying user password hash _(openshift 4.13.0+)_
 - Support offline Tang provisioning via pre-shared advertisement _(fcos 1.5.0-exp, openshift 4.14.0-exp)_
+- Support local file embedding for SSH keys and systemd units
 
 ### Bug fixes